requirements.txt
colorama~=0.4.6
docopt~=0.6.2
pandas~=2.2.2
Pillow~=10.4.0
requests~=2.32.3
smart-open~=7.0.4
tabulate~=0.9.0
tqdm~=4.66.4
setuptools~=70.3.0